% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHNRegionCode.R
\name{CHNRegionCode}
\alias{CHNRegionCode}
\title{Title}
\usage{
CHNRegionCode(import, year_col, from_year, to_year = 2015, type)
}
\arguments{
\item{import}{A character vector}

\item{year_col}{A character vector}

\item{from_year}{A character vector}

\item{to_year}{A character vector}

\item{type}{A character vector}
}
\description{
Title
}
