% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corruption.R
\docType{data}
\name{corruption}
\alias{corruption}
\title{\link{Filling the Title}}
\format{
A data frame with XXXXXX10 variables:
\describe{
\item{name}{description}
\item{name}{description}
...
}
}
\source{
\url{http://www.diamondse.info/} # link where you got the data if there is
}
\usage{
corruption
}
\description{
A dataset containing \link{Filling the description}
}
\keyword{datasets}
