% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHNRegionCode.R
\name{CHNRegionCode}
\alias{CHNRegionCode}
\title{CHNRegionCode}
\usage{
CHNRegionCode(import, year_col, from_year, to_year = 2015, type)
}
\arguments{
\item{import}{Table virable that needs geocode coverting.The class of import virable must be dataframe.}

\item{year_col}{Column that needs to be coverted.}

\item{from_year}{Code, name or sname version of \code{year_col}.}

\item{to_year}{Code, name or sname version you want get. The default is 2015.}

\item{type}{The function converting type. The Valid strings are: \code{name2name}, \code{name2sname}, \code{name2code}, \code{code2code}, \code{code2name}, \code{code2sname}, \code{sname2sname}, \code{sname2name}, and \code{sname2code}.}
}
\description{
\code{regioncode} is an software developes to conquer the difficulties to match Chinese regional data across years due to such geocodes adjustments. In the current version, \code{regioncode} enables seamlessly converting Chinese regions' formal names, common-used names, and geocodes between each other and across thirty-four years from 1986.
}
\examples{
# library(regioncode)
# load("./vignettes/vignette_data.rda")
# vignette_data <- as.data.frame(vignette_data)
# vignette_covert <- CHNregioncode(import = vignette_data,
#                    year_col = 'prefecture_id',from_year = 2019,to_year = 2005,type = "code2code")

}
